requirements.txt
# Core Application Dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.4.0
openai==1.43.0
httpx==0.27.0  # Required: OpenAI 1.43.0 is incompatible with httpx 0.28+ (proxies vs proxy)
beautifulsoup4==4.12.3
lxml==5.1.0
PyPDF2==3.0.1
pdfplumber==0.11.1
python-dotenv==1.0.1
aiofiles==24.1.0
requests==2.31.0  # For Streamlit frontend to call backend API
streamlit==1.38.0
sec-edgar-downloader==5.0.0
pyrate-limiter==3.1.0

# Yahoo Finance data source
yfinance>=0.2.28

# Development and Testing Dependencies
pytest>=7.4.0,<8.0
pytest-cov>=4.1.0,<5.0
pytest-mock>=3.12.0,<4.0
mypy>=1.8.0,<2.0
black>=24.1.0,<25.0
ruff>=0.5.0,<1.0
